backend/requirements.txt
aioredis==1.3.1
amqp==5.0.9
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.1
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-redis-cache==3.0.0
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.3
msgpack==1.0.3
promise==2.3
prompt-toolkit==3.0.24
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
requests==2.27.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
zope.interface==5.4.0
